se implementa react router para separar las rutas
Esto se hace a futuro para separar los modulos de Mandala app
